Mom's wish is my command :)
My mother asked me whether it was possible to show the number of moves
during a cardgame, "just like in Windows". It's possible now.

The statusbar will show how many moves are made during a game. It's
currently tracked for each gametype for now, I'll try to fix this if
there are gametypes where move-count doesn't make sense.

svn path=/trunk/kdegames/kpat/; revision=311711