Fix bug 87451: crash in patience when selecting 'unknown' as game type
 - renumber the following patiences: Golf, Klondike (draw 3), Spider
   Easy, Spider Medium, Spider Hard.

This bug was due to the fact that Stephan Kulow wasn't thorough enough
when he disabled the Kings patience because of bug 84819.

Well, now it seems to be really disabled.  Now all that remains is to
actually *fix* 84819. :-)

I will find out if this fix needs backporting.

CCMAIL: 84819@bugs.kde.org
CCMAIL: 87451-done@bugs.kde.org

svn path=/trunk/kdegames/kpat/; revision=354544