Update the statistics all in one spot, and only at the end of the game or when abandoning it for a new one. This should work out most of the kinks related to saving state between sessions and game stats.
Previously, the game was counted toward the total of games played at the start of the game, but the win or loss was counted at the end of the game. This meant that if KPat crashed midgame, it would be counted as a loss. It also allowed a loophole through which you could start a game then save it, reopen it, beat it, reopen it, beat it, etc. The game was started only once but won multiple times, so the win precentage could easily exceed 100%.

svn path=/trunk/KDE/kdegames/kpat/; revision=936043