Switched from KInputDialog::getText() to KInputDialog::getInteger() for the Choose Game... action. Also cleaned up the internal handling of game numbers.
Certain parts of the code used ints while others used longs. This is a non issue on most 32bit system where ints and longs are the same size. However, on 64bit systems where they differ, it gave a false impression of range. Certain parts of the code are restricted to ints, so passing around longs was only confusing.

If there are any 64bit users out there who have already beaten the first 2147483647 games of Klondike, let me know and I'll consider upgrading the code to qint64.

svn path=/trunk/KDE/kdegames/kpat/; revision=982896