Modified how spaces are inserted between conflicting piles.
Formerly, the maximumSpaces of conflicting piles would butt up against one another. However the cards would never actually touch, because the pile layout took into account the spread on the top card. As a result, there would always be a gap of spread between the top card and the end of the maximumPosition. Where this is a proble is where spread is large, or when a pile touches the bottom of the contents area. It seems weird to have a pile stop short of the bottom of the screen for no good reason.

Now, when handling pile conflicts, a space (currently 1.5) is placed between the piles and the pile layout algorithm ignores the spread on the top card. Everybody wins! ;)

svn path=/trunk/KDE/kdegames/kpat/; revision=1010654