Further clean up of class interdependencies.
DealerScene subclasses no longer find the MainWindow instance and insert their actions into the menu themselves. Instead they have a getter that returns the list, which the mainwindow then inserts itself.

As a result, View no longer needs to keep track of its instance or the instance of the MainWindow that contains it. This means even more code removed from View.

svn path=/trunk/KDE/kdegames/kpat/; revision=1012201