Fixed a bug where the title bar didn't relfect the new game number when changing Klondike or Spider options.
Previously, when changing the options in Klondike or Spider, a new random game was opened, but the game number shown in the titlebar wasn't updated. Getting this to work correctly would be tricky, so I took the lazy route and made it just reload the same game number. No need to update the titlebar!

Backport of 1030625.

svn path=/branches/KDE/4.3/kdegames/kpat/; revision=1030628