Changed the Card transformationMode to SmoothTransformation.
This corrects and issue that has been showing up in recent KPat trunk. When right clicking on a card to make it "pop up" for inspection the card graphics became very grainy with some obvious scaling artifacts. I haven't yet deduce what caused this change, but this fixes it. SmoothTransform is more expensive, but the Cards are transformed so rarely, I don't think this is an issue.

svn path=/trunk/KDE/kdegames/kpat/; revision=1032480