Prevent the solver timer from being started before the cards have been dealt.
Previously the solver timer was being started when the DealerScene was first created. I think this was causing crashes because occasionally the timer would time out and the solver would fire up before all the cards had actually been dealt. The solver code is very lean and does no error checking, so it would just crash if it expected cards and didn't get them.

svn path=/trunk/KDE/kdegames/kpat/; revision=1036102