Drastically simplified the painting of Cards.
The following is taken from the ReviewBoard review request:
Lately I've been trying to clean up KPat codebase and have had a bit of success. Yesterday I had a read through the Card class and noticed that it contained lots of painting optimisations.

- Each Card often checks to see if it is obscured by other cards.
- Each Card has stores whether it is visible, hidden, or doesn't know.
- Each Card keeps a list of cards it obscures and rechecks their visibility when it moves.
- Only the exposed portions of a card are repainted.

I found this quite surprising as Card is a subclass of QGraphicsPixmapItem and I was under the impression that the GraphicsView framework handled a lot of these things internally. I did some testing an found that this code wasn't really working as expected. If 52 cards are pile directly on top of one another, dragging a card over top of them caused all 52 of them to repaint, even though only the top one is actually visible.

Of course KPat was one of the first KDE apps ported to Qt4 and GraphicsView. GraphicsView was brand new then and maybe these optimisations were needed to help it along. Since then things have changed quite a bit which may have internally broken these optimisations. This patch radically simplifies the painting logic.

svn path=/trunk/KDE/kdegames/kpat/; revision=1039741