Removed the "getUp" functionality.
Previously, if you right clicked on a face down card, for some (intentional?) reason it would jump to the top for 1.5 seconds. I have no idea why it did this, but it always seemed broken to me.

Now right clicking on a facedown card will use the same zoom/slide/rotate animation as face up cards. Is this useful? Not really, but who am I to tell which cards you can zoom/slide/rotate and which you can't.

svn path=/trunk/KDE/kdegames/kpat/; revision=1040150