Add separate solver texts for games that were never winnable.
Previously, whenever the game was unwinnable, the status bar showed
"Solver: This game is not winnable in its current state." which is a
little bit confusing. Now it shows "Solver: This game cannot be won."
if the game has never been winnable and "Solver: This game is no longer
winnable." if it was winnable, but through the player's actions becomes
unwinnable.

svn path=/trunk/KDE/kdegames/kpat/; revision=1050757