Add a signal to CardDeck to notify when all animations have stopped.
This provides an easy way to delay some action until after the current
animations have stopped. I think this has potential to replace some
rather ugly code in DealerScene that currently relies on starting a
timer connected to a slot, where the slot continues to restart the timer
as long as animations are going. Using this signal is prettier, gives
quicker responses and is more efficient.

svn path=/trunk/KDE/kdegames/kpat/; revision=1057559