Simplify CardDeck's animated Card tracking.
Originally CardDeck tracked the number of cards currently being
animated by storing pointers to said Cards in a QSet. This made
debugging easier and and allowed us to check that a single card
wasn't added or removed more than in a row.

Now that the code has been working for while without issue, it makes
sense to replace the QSet with a simple int that is incremented and
decremented.

svn path=/trunk/KDE/kdegames/kpat/; revision=1057660