Work around bug 215136.
The demo mode was getting stuck, because calling translateMove on the
first element of minMoves returned 0. In such a case, I think it makes
sense to fallback on one of the other hints, if present.

I'm not closing the bug, as this merely patches over the root problem.
Obviously if I move is the list of winning moves, it should be valid
and translatable.

CCBUG: 215136

svn path=/trunk/KDE/kdegames/kpat/; revision=1057664