Remove Patience specific flags from Pile.
This is a major change. Instead of setting lots of magical KPat
specific flags in Pile and using those to sometimes check if Cards are
allowed to be added or removed, these checks are now handled in the
CardScene subclasses where they belong. checkAdd() and checkRemove() are
now much more important now, since all card adds and removes pass
through them. Still more cleanup to come.

svn path=/trunk/KDE/kdegames/kpat/; revision=1075368