A workaround for bug 215136.
The demo mode was getting stuck, because I had prevented moving a
single card back from the waste to the stock (since it was pointless),
but the solver was still suggesting moves of this type.

As a workaround for SC4.4, I'm allowing such pointless cycling from
the stock to the waste and back. I'm not closing the bug however, as
I'd like to keep the check and make the appropriate change to the
solver for SC4.5.

CCBUG:215136

svn path=/branches/KDE/4.4/kdegames/kpat/; revision=1075811