Add KConfigXT support to KCardThemeWidget.
This has a few benefits. The biggest is that the user can now use the
Apply button to check out the look of a theme before closing the dialog.
It also means the selection dialog doesn't have to be modal any more. Of
course, since it's KConfigXT we are able to drop some more manual config
reading and writing code.

svn path=/trunk/KDE/kdegames/kpat/; revision=1086727