Add new private KCardAnimation to replace complex QPropertyAnimations.
Rather than set up a bunch of properties on KCardPrivate with a bunch of
pass through methods, just to connect to a QParallelAnimationGroup of
QPropertyAnimations, it's much cleaner, simpler and less code to
implement a custom animation class. Because KCardAnimation is so
specialised, it has very little code and may even be more performant.

svn path=/trunk/KDE/kdegames/kpat/; revision=1086741