Remove the concept of "undealt cards" in KAbstractCardDeck.
Long ago, Deck used to be a specialised type of Pile that had extra
methods for shuffling and things. It's partner CardMap handled the
resizing and pixmaps of the Cards. Deck being a Pile was a bit
awkward as it had to belong to the DealerScene and games that didn't
use a stock had an extra usused pile.

I then merged Deck and CardMap into a single class called CardDeck.
CardDeck became the single source for most deck related options
including shuffling. It also added the concept of undealt cards,
initially all cards were undealt, and the DealerScene would then
"take" cards and add them to the scene or "return" cards to join the
ranks of the undealt. These were mostly holdovers from the fact that
deck used to be an actual Pile and have ownership of some cards.

Nowadays, I don't think the concept of undealt cards makes much sense.
They were basically only being used during the initial deal. CardScenes
that need a concept of undealt cards are probably going to want
something more advanced than what KAbstractCardDeck was offering anyway.
For initial deals, the DealerScene subclasses just grab a copy of the
KAbstractCardDeck's card list and then remove cards from it as they're
dealt.

I think this really improves the API. KAbstractCardDeck is now dedicated
to the creation, ownership and updating up KCards and not game specific
details.

svn path=/trunk/KDE/kdegames/kpat/; revision=1092016