Add delayed KCard rendering.
This is a feature I've been wanting to add for a while. When the theme
or card size changes, the card pixmaps are quickly (and crudely) scaled
to the correct size. A rendering thread is then started and the cards
are updated as new pixmaps arrive.

This makes things much more responsive during resize. There are plenty
of further improvements and optimisations that I'm looking into, but
this is the first working version, so I thought it made sense to commit
it.

svn path=/trunk/KDE/kdegames/kpat/; revision=1092018