Make flip animation last only over the initial portion of the movement.
Now when flipping a card over a long distance, now only the initial part
of the travel is spent flipping the card and the card remains fully
flipped for the remainder of the travel. This looks a bit more
realistic.

Also optimised KCardAnimation progress calculations by storing the
deltas instead of the final values. This saves a substraction per
property on each update.

svn path=/trunk/KDE/kdegames/kpat/; revision=1094084