Don't remember state on exit if game has been won.
Previously, if Remember State on Exit was enabled and KPat was saved
while on the "Congratulations" screen, KPat would save an almost empty
state file which would just launch a new game of that same type on
next startup. I'm not sure that this is actually a good idea, because
on next startup the user may confuse the new game as a game that they
have already started.

Now, we don't bother saving the state to file when on the congrats
screen, so that the game selection screen is shown on next startup.

svn path=/trunk/KDE/kdegames/kpat/; revision=1122776