Remove the solver mutex from DealerScene.
Previously, DealerScene did quite a bit of manual mutex locking and
unlocking. If one wasn't careful when changing the surrounding code, one
could easily cause QMutex asserts. Tonight, after one such incident, I
put in a genuine effort to understand how this locking code worked.

It turns out that this code was completely unneeded. The mutex is only
ever used from the main thread and it doesn't protect and thread mutable
data. All of it could be removed and replaced with a few calls to
m_solver_thread->isRunning(). Presumably this code was needed for thread
safety in the past, but now we can just get rid of it entirely.

There may be some edges cases that this isn't handling perfectly, but
we'll flush them out if they arise. A small price to pay to avoid all
those mutex related crashes, if you ask me. :D

svn path=/trunk/KDE/kdegames/kpat/; revision=1122779