Create a system for queuing DealerScene methods until animation is done.
There are several methods in DealerScene that if called during an
animation have to be delayed until after the animation has completed.
Previously, this was implemented by each method having it's own QTimer,
which it would use to call itself after a certain delay if it detected
animation. This wasn't great for serveral reasons: it was inefficient,
it was ugly, and in unfortunate cases, new user imput would come (or
some other timer would expire) in between the end of the animation and
the expiration of the timer which resulted in methods being called out'
of sequence.

Since KPat3.3, we've had a signal to notify us when all card animations
finished. This commit does the obvious thing and replaces the timers
with a simple queuing system. If a method attempts to run while an
animation is running, it just sets a flag. When the animations finish,
the animationDone slot is called, which checks the flags and calls the
corresponding methods. This could still probably be cleaner, but it's a
major improvement in my eyes.

svn path=/trunk/KDE/kdegames/kpat/; revision=1122784