Mostly revert commit 1132430.
This is a backport of commit 1156130.

In said commit I decided to be clever and not "call stop() on mouse
release or double-click events", arguing "We don't get a release or
double-click event without first getting a press event, so calling
stop() in mousePressEvent() is sufficient."

In fact, I was dead wrong. We do indeed always get a press event, but
there's nothing stopping an action from starting in between the mouse
press and the subsequent release. While this is generally unlikely, that
is exactly what was occuring in bug 245825.

The first right button release started the card moving to the
foundation pile. The second right button press occured while the first
card was in transit. The first card then reached its destination pile,
triggering an autodrop to begin. The second right button release then
started to tryAutomaticMove(), which it did successfully, but because
there was still an autodrop in progreess resulting state wasn't being
properly handled.

This was a doozy to track down, but downright trivial to fix.

CCBUG:245825

svn path=/branches/KDE/4.5/kdegames/kpat/; revision=1156133