Add initial support for sound effects in KPat.
This commit adds simple Phonon-based sound effects to KPat. At the
moment sounds are only emitted when cards are picked up or set down,
but hopefully more sounds can be added in the future.

Because Phonon doesn't support playing multiple concurrent sounds,
sounds are sometimes cut short. I'm not sure if there much that can be
done about that, though.

svn path=/trunk/KDE/kdegames/kpat/; revision=1166075