Correct card ordering in KCardScene::flipCardsToPile().
When flipping multiple cards from one pile to another, we have to
reverse the stacking order of the cards. The topmost card on the source
pile becomes the bottommost card on the destination pile. This fixes
the waste to stock recycling in both Klondike and Forty & Eight.

The previous behaviour wasn't necessarily wrong, but it did represent a
maneuver that can't really be done with real cards in a single move,
hence it wasn't the correct behaviour for the games in question.

svn path=/trunk/KDE/kdegames/kpat/; revision=1229903