Remove a bunch of cruft from DealerScene::demo().
This code was doing a bunch of things suboptimally. It didn't use
KCardPile::topCardsDownTo(). It handled a bunch of animation stuff
that should now be handled automatically. It asserted pointers after
having used them, which is pointless.

While I don't know what was causing the crash in bug 274127, I do know
it can't happen again with the new code.

BUG:274127

svn path=/trunk/KDE/kdegames/kpat/; revision=1262923