Backport of commit 1275129.
Don't bother saving state on close if the player has already won.

This was a behaviour regression caused by the clean up of the loading
and saving code. Somewhat surprisingly, saving after the game has been
won worked without issue and loading that save caused the player to
instantly win again, so it acted like the state was actually remembered.
Of course, the user doesn't expect to fire up KPat just to be instantly
told that they've won.

CCBUG:291959

svn path=/branches/KDE/4.8/kdegames/kpat/; revision=1275131