Make KCardDeck card IDs unique, even for cards with same suit and rank.
Previously, two KCards having the same suit and rank would have
identical IDs, leaving no real way (other than their addresses) to tell
them apart at run time. This makes implementing some changes I have
planned a real hassle.

Since KCardDeck was only using the lowest 6 bits of the 32 available in
the ID, we have plenty of room to store the card's position within the
deck. The new layout is 8 bits for the rank, 8 for the suit, and the
remaining 16 for the number. So that's support for 256 different ranks,
256 different suits and 65536 different cards.

This change is also made to KPat's slightly custom ID generation code.

svn path=/trunk/KDE/kdegames/kpat/; revision=1280163