Add initial support for saving/loading history to/from file.
Instead of saving the just current state of the game, KPat now saves
the entire history, allowing one to undo/redo in the loaded file. I've
been wanting to do this for a while and finally got it finished.

This is a brand new and entirely separate XML format. KPat can open
both kinds of files and when saving the user can choose either one,
although we default to the new format. The old format will stay in
place for at least a few releases, but for simplicity's sake we might
want to remove support for at some point.

Hopefully it'll improve the quality of bug reports, since the user will
be able to attach the exact series of moves that lead up to a problem.
It also means that people won't lose their undo history when closing
KPat with "Remember State on Exit" checked.

In addition to the new history stuff, I've tried to improve the file
format to make it more human readable and therefore, hopefully more
useful for debugging.

I still have a few changes I want to make to the code and to the save
file format, but this is a decent start that seems to work quite well
already.

DIGEST:

svn path=/trunk/KDE/kdegames/kpat/; revision=1280168