Allow autodropping only when the redo stack is empty.
If there are states in the redo stack, there's a pretty good chance
that the user would like to keep them, but autodropping would replace
them with a trivial move. This isn't an issue in normal play, but can
come up when the user plays with autodrop disabled, misses some drops,
enables autodrop and then undoes back a few states.

svn path=/trunk/KDE/kdegames/kpat/; revision=1280178