Reduce wasted space for Klondike and Grandfather
Summary:
Klondike and (especially) Grandfather reserves a lot of
space at the bottom of the screen for long card piles.
This results in very small cards, which makes the game
uncomfortable to play, especially on smaller screens.

This commmit reduces the space reserved at the bottom
of the screen, which automatically enlarges the cards.
Long piles don’t pose a problem, as KPat automatically
compresses the cards to fit the screen.

Even with the reduced space, the colours and numbers are
still clearly visible, also in a ‘worst case’ scenario with very
long piles. (In these two game types, only the colour
(i.e. red vs. black) matter, not the suit, so it is not a
problem if the suits of the cards aren’t visible.)

BUG: 307625
BUG: 378914

FIXED-IN: 18.08.2

Test Plan:
Play Grandfather and Klondike before and after this patch.
Note that the cards are much bigger after the patch
and that the numbers and colours are still clearly visible.

Reviewers: #kde_games, fabiank, mlaurent

Reviewed By: fabiank

Subscribers: kde-games-devel

Differential Revision: https://phabricator.kde.org/D15534