Add a comment to explain.
Sometimes the solver is invoked with a small maximal iterations
quota/limit (e.g: when loading a saved game or checking for autodrop
moves) and it is done frequently, so we return prematurely without
invoking freecell_solver_user_alloc() and friends which incur extra
overhead.

The m_firstMoves should be good enough in that case.