Add minimal number of moves statistics
Summary:
Statistics that displays the minimal number of moves that a player used to win a game. This could be used to make the game more competitive - comparing your score to other people's.

The number is set to -1 (which is unachievable) when reseting the statistics and an infinity symbol is then displayed.

Reviewers: #kde_games

Subscribers: aacid, kde-games-devel

Differential Revision: https://phabricator.kde.org/D25642