PoC: show how greenblaze terribly impacts performance
The greenblaze theme currently use a lot of embedded images. There is one for
the background, obviously, but each item uses one, for an object that seems
quite easy to do in pure SVG: a rectangle with a gradient.
In this commit, I removed all these images, keeping only the background.
Since I have absolutely no SVG nor inkscape skills, I could not replace them
with proper objects. But the impact on the theme is almost non-existing, while
the performance impact is surprising: startup time goes from 1.6 to 0.3 seconds
on my computer with this change.