Add three suit game mode to Spider
This change probably isn't necessary, however it would be nice to have
a difficulty level higher than two suits that isn't quite as brutal as
four suits.

Add the necessary logic for a deck of spades, hearts, and clubs.
Name this new difficulty level of three suit to "Hard".
Rename the difficulty level of four suit to "Very Hard" (previously
"Hard").