Fix simple simon auto drop: always use full hints if solver has 0
The way the fc-solve backend is used doesn't provide win moves at
the time the auto drop is calculated - or first moves at all. So
we can't use it for this and fallback to brute force method we used
before solvers were a thing.

This makes the hint a little more spammy, but at least auto drop
behaves as expected

Fixes bug 411273