Fix spider solver to be valgrind free
The actual crash was caused by a rather optimistic use of magic
numbers. But I took the oportunity to remove some more code smells
and made sure valgrind no longer complains about anything (also
removed global variable that was fine when this was a standalone
program to experiment with, but no longer as part of a background
thread).

BUG 466229