Solve the simple simon autodrop problem less aggressively
Using the brute force algorithm for auto drop has several drawbacks
that users reported:
  - it's too aggressive (didn't matter for simple simon when I implemented
    it, as there is only one kind of foundation move)
  - the hint function is no longer a way to cheat your scores, you actually
    need to pick the winning move yourself

So revert that part and make in return the simple simon solver more useful
for autodrop by providing the right priority for foundation moves. In addition
I had to make sure we always return a first move even if the old interface
is not implemented.

BUGS: 469991